- Adding method to collect refused episodes for report - Adding SQL to power the method - Adding tests for method
