+++ nh_eobs_mental_health/sql/__init__.py
@@ -1,0 +1,2 @@
+from . import ward_dashboard
+from . import refused_observations
