+++ nh_eobs_mental_health/tests/report_nh_clinical_observation_report/__init__.py
@@ -1,1 +1,2 @@
 from . import test_get_report_data
+from . import test_get_refusal_episodes
